// src/stats/types/team-dashboard-by-general-splits.ts
export interface TeamDashboardByGeneralSplitsParams {
	teamID: number;
	lastNGames?: number;
	measureType?: string;
	month?: number;
	opponentTeamID?: number;
	paceAdjust?: string;
	perMode?: string;
	period?: number;
	plusMinus?: string;
	rank?: string;
	season?: string;
	seasonType?: string;
	dateFrom?: string;
	dateTo?: string;
	gameSegment?: string;
	leagueID?: string;
	location?: string;
	outcome?: string;
	pORound?: number;
	seasonSegment?: string;
	shotClockRange?: number;
	vsConference?: string;
	vsDivision?: string;
}

export interface DaysRestTeamDashboardRow {
	groupSet: unknown;
	groupValue: unknown;
	teamDaysRestRange: unknown;
	gp: number;
	w: number;
	l: number;
	wPct: number | null;
	min: number;
	fgm: number;
	fga: number;
	fgPct: number | null;
	fg3m: number;
	fg3a: number;
	fg3Pct: number | null;
	ftm: number;
	fta: number;
	ftPct: number | null;
	oreb: number;
	dreb: number;
	reb: number;
	ast: number;
	tov: number;
	stl: number;
	blk: number;
	blka: number;
	pf: number;
	pfd: number;
	pts: number;
	plusMinus: number;
	gpRank: number;
	wRank: number;
	lRank: number;
	wPctRank: number;
	minRank: number;
	fgmRank: number;
	fgaRank: number;
	fgPctRank: number;
	fg3mRank: number;
	fg3aRank: number;
	fg3PctRank: number;
	ftmRank: number;
	ftaRank: number;
	ftPctRank: number;
	orebRank: number;
	drebRank: number;
	rebRank: number;
	astRank: number;
	tovRank: number;
	stlRank: number;
	blkRank: number;
	blkaRank: number;
	pfRank: number;
	pfdRank: number;
	ptsRank: number;
	plusMinusRank: number;
	cfid: number;
	cfparams: unknown;
}

export interface LocationTeamDashboardRow {
	groupSet: unknown;
	groupValue: unknown;
	teamGameLocation: unknown;
	gp: number;
	w: number;
	l: number;
	wPct: number | null;
	min: number;
	fgm: number;
	fga: number;
	fgPct: number | null;
	fg3m: number;
	fg3a: number;
	fg3Pct: number | null;
	ftm: number;
	fta: number;
	ftPct: number | null;
	oreb: number;
	dreb: number;
	reb: number;
	ast: number;
	tov: number;
	stl: number;
	blk: number;
	blka: number;
	pf: number;
	pfd: number;
	pts: number;
	plusMinus: number;
	gpRank: number;
	wRank: number;
	lRank: number;
	wPctRank: number;
	minRank: number;
	fgmRank: number;
	fgaRank: number;
	fgPctRank: number;
	fg3mRank: number;
	fg3aRank: number;
	fg3PctRank: number;
	ftmRank: number;
	ftaRank: number;
	ftPctRank: number;
	orebRank: number;
	drebRank: number;
	rebRank: number;
	astRank: number;
	tovRank: number;
	stlRank: number;
	blkRank: number;
	blkaRank: number;
	pfRank: number;
	pfdRank: number;
	ptsRank: number;
	plusMinusRank: number;
	cfid: number;
	cfparams: unknown;
}

export interface MonthTeamDashboardRow {
	groupSet: unknown;
	groupValue: unknown;
	seasonMonthName: string;
	gp: number;
	w: number;
	l: number;
	wPct: number | null;
	min: number;
	fgm: number;
	fga: number;
	fgPct: number | null;
	fg3m: number;
	fg3a: number;
	fg3Pct: number | null;
	ftm: number;
	fta: number;
	ftPct: number | null;
	oreb: number;
	dreb: number;
	reb: number;
	ast: number;
	tov: number;
	stl: number;
	blk: number;
	blka: number;
	pf: number;
	pfd: number;
	pts: number;
	plusMinus: number;
	gpRank: number;
	wRank: number;
	lRank: number;
	wPctRank: number;
	minRank: number;
	fgmRank: number;
	fgaRank: number;
	fgPctRank: number;
	fg3mRank: number;
	fg3aRank: number;
	fg3PctRank: number;
	ftmRank: number;
	ftaRank: number;
	ftPctRank: number;
	orebRank: number;
	drebRank: number;
	rebRank: number;
	astRank: number;
	tovRank: number;
	stlRank: number;
	blkRank: number;
	blkaRank: number;
	pfRank: number;
	pfdRank: number;
	ptsRank: number;
	plusMinusRank: number;
	cfid: number;
	cfparams: unknown;
}

export interface OverallTeamDashboardRow {
	groupSet: unknown;
	groupValue: unknown;
	seasonYear: string;
	gp: number;
	w: number;
	l: number;
	wPct: number | null;
	min: number;
	fgm: number;
	fga: number;
	fgPct: number | null;
	fg3m: number;
	fg3a: number;
	fg3Pct: number | null;
	ftm: number;
	fta: number;
	ftPct: number | null;
	oreb: number;
	dreb: number;
	reb: number;
	ast: number;
	tov: number;
	stl: number;
	blk: number;
	blka: number;
	pf: number;
	pfd: number;
	pts: number;
	plusMinus: number;
	gpRank: number;
	wRank: number;
	lRank: number;
	wPctRank: number;
	minRank: number;
	fgmRank: number;
	fgaRank: number;
	fgPctRank: number;
	fg3mRank: number;
	fg3aRank: number;
	fg3PctRank: number;
	ftmRank: number;
	ftaRank: number;
	ftPctRank: number;
	orebRank: number;
	drebRank: number;
	rebRank: number;
	astRank: number;
	tovRank: number;
	stlRank: number;
	blkRank: number;
	blkaRank: number;
	pfRank: number;
	pfdRank: number;
	ptsRank: number;
	plusMinusRank: number;
	cfid: number;
	cfparams: unknown;
}

export interface PrePostAllStarTeamDashboardRow {
	groupSet: unknown;
	groupValue: unknown;
	seasonSegment: string;
	gp: number;
	w: number;
	l: number;
	wPct: number | null;
	min: number;
	fgm: number;
	fga: number;
	fgPct: number | null;
	fg3m: number;
	fg3a: number;
	fg3Pct: number | null;
	ftm: number;
	fta: number;
	ftPct: number | null;
	oreb: number;
	dreb: number;
	reb: number;
	ast: number;
	tov: number;
	stl: number;
	blk: number;
	blka: number;
	pf: number;
	pfd: number;
	pts: number;
	plusMinus: number;
	gpRank: number;
	wRank: number;
	lRank: number;
	wPctRank: number;
	minRank: number;
	fgmRank: number;
	fgaRank: number;
	fgPctRank: number;
	fg3mRank: number;
	fg3aRank: number;
	fg3PctRank: number;
	ftmRank: number;
	ftaRank: number;
	ftPctRank: number;
	orebRank: number;
	drebRank: number;
	rebRank: number;
	astRank: number;
	tovRank: number;
	stlRank: number;
	blkRank: number;
	blkaRank: number;
	pfRank: number;
	pfdRank: number;
	ptsRank: number;
	plusMinusRank: number;
	cfid: number;
	cfparams: unknown;
}

export interface WinsLossesTeamDashboardRow {
	groupSet: unknown;
	groupValue: unknown;
	gameResult: unknown;
	gp: number;
	w: number;
	l: number;
	wPct: number | null;
	min: number;
	fgm: number;
	fga: number;
	fgPct: number | null;
	fg3m: number;
	fg3a: number;
	fg3Pct: number | null;
	ftm: number;
	fta: number;
	ftPct: number | null;
	oreb: number;
	dreb: number;
	reb: number;
	ast: number;
	tov: number;
	stl: number;
	blk: number;
	blka: number;
	pf: number;
	pfd: number;
	pts: number;
	plusMinus: number;
	gpRank: number;
	wRank: number;
	lRank: number;
	wPctRank: number;
	minRank: number;
	fgmRank: number;
	fgaRank: number;
	fgPctRank: number;
	fg3mRank: number;
	fg3aRank: number;
	fg3PctRank: number;
	ftmRank: number;
	ftaRank: number;
	ftPctRank: number;
	orebRank: number;
	drebRank: number;
	rebRank: number;
	astRank: number;
	tovRank: number;
	stlRank: number;
	blkRank: number;
	blkaRank: number;
	pfRank: number;
	pfdRank: number;
	ptsRank: number;
	plusMinusRank: number;
	cfid: number;
	cfparams: unknown;
}

export interface TeamDashboardByGeneralSplitsResponse {
	daysRestTeamDashboard: DaysRestTeamDashboardRow[];
	locationTeamDashboard: LocationTeamDashboardRow[];
	monthTeamDashboard: MonthTeamDashboardRow[];
	overallTeamDashboard: OverallTeamDashboardRow[];
	prePostAllStarTeamDashboard: PrePostAllStarTeamDashboardRow[];
	winsLossesTeamDashboard: WinsLossesTeamDashboardRow[];
}